+++ src/storage/source.rs
@@ -1,10 +1,10 @@
 use {
     crate::{
-        config::ConfigSource,
+        config::{ConfigSource, ConfigSourceStream},
         source::{
             block::BlockWithBinary,
             rpc::{GetBlockError, RpcSource},
-            stream::{RecvError, StreamSource, StreamSourceMessage},
+            stream::{StreamSource, StreamSourceMessage},
         },
     },
     futures::stream::StreamExt,
@@ -13,14 +13,17 @@ use {
     solana_sdk::clock::Slot,
     std::sync::Arc,
     thiserror::Error,
-    tokio::sync::{Notify, mpsc, oneshot},
+    tokio::{
+        sync::{Notify, mpsc, oneshot},
+        time::sleep,
+    },
     tracing::error,
 };
 
 #[derive(Debug)]
 pub enum RpcRequest {
-    ConfirmedSlot {
-        tx: oneshot::Sender<Result<Slot, ClientError>>,
+    Slots {
+        tx: oneshot::Sender<Result<(Slot, Slot), ClientError>>,
     },
     Block {
         slot: Slot,
@@ -66,9 +69,9 @@ impl RpcSourceConnected {
         }
     }
 
-    pub async fn get_slot_confirmed(&self) -> RpcSourceConnectedResult<Slot, ClientError> {
+    pub async fn get_slots(&self) -> RpcSourceConnectedResult<(Slot, Slot), ClientError> {
         let (tx, rx) = oneshot::channel();
-        self.send(RpcRequest::ConfirmedSlot { tx }, rx).await
+        self.send(RpcRequest::Slots { tx }, rx).await
     }
 
     pub async fn get_block(
@@ -88,31 +91,78 @@ pub async fn start(
     shutdown: Shutdown,
 ) -> anyhow::Result<()> {
     let rpc = Arc::new(RpcSource::new(config.rpc).await?);
-    let mut stream: Option<StreamSource> = None;
+    let stream = start_stream(config.stream, stream_tx, stream_start);
 
     tokio::pin!(shutdown);
+    tokio::pin!(stream);
+
     let mut finished = false;
     while !finished {
-        finished = if let Some(stream) = stream.as_mut() {
-            tokio::select! {
-                () = &mut shutdown => true,
-                item = rpc_rx.recv() => handle_rpc(item, &rpc),
-                item = stream.next() => handle_stream(item, &stream_tx).await
-            }
-        } else {
-            tokio::select! {
-                () = &mut shutdown => true,
-                item = rpc_rx.recv() => handle_rpc(item, &rpc),
-                () = stream_start.notified() => {
-                    stream = Some(StreamSource::new(config.stream.clone()).await?);
-                    false
-                },
-            }
+        finished = tokio::select! {
+            () = &mut shutdown => true,
+            item = rpc_rx.recv() => handle_rpc(item, &rpc),
+            result = &mut stream => return result,
         };
     }
     shutdown.shutdown();
 
     Ok(())
+}
+
+async fn start_stream(
+    config: ConfigSourceStream,
+    stream_tx: mpsc::Sender<StreamSourceMessage>,
+    stream_start: Arc<Notify>,
+) -> anyhow::Result<()> {
+    let mut backoff_duration = config.reconnect.map(|c| c.backoff_max);
+    let backoff_max = config.reconnect.map(|c| c.backoff_max).unwrap_or_default();
+
+    stream_start.notified().await;
+    loop {
+        let mut stream = loop {
+            match StreamSource::new(config.clone()).await {
+                Ok(stream) => break stream,
+                Err(error) => {
+                    if let Some(sleep_duration) = backoff_duration {
+                        error!(?error, "failed to connect to gRPC stream");
+                        sleep(sleep_duration).await;
+                        backoff_duration = Some(backoff_max.min(sleep_duration * 2));
+                    } else {
+                        return Err(error.into());
+                    }
+                }
+            }
+        };
+        if stream_tx.send(StreamSourceMessage::Start).await.is_err() {
+            error!("failed to send a message to the stream");
+            return Ok(());
+        }
+
+        loop {
+            match stream.next().await {
+                Some(Ok(message)) => {
+                    if stream_tx.send(message).await.is_err() {
+                        error!("failed to send a message to the stream");
+                        return Ok(());
+                    }
+                }
+                Some(Err(error)) => {
+                    error!(?error, "gRPC stream error");
+                    break;
+                }
+                None => {
+                    error!("gRPC stream is finished");
+                    break;
+                }
+            }
+        }
+
+        if let Some(config) = config.reconnect {
+            backoff_duration = Some(config.backoff_init);
+        } else {
+            return Ok(());
+        }
+    }
 }
 
 fn handle_rpc(item: Option<RpcRequest>, rpc: &Arc<RpcSource>) -> bool {
@@ -121,8 +171,9 @@ fn handle_rpc(item: Option<RpcRequest>, rpc: &Arc<RpcSource>) -> bool {
             let rpc = Arc::clone(rpc);
             tokio::spawn(async move {
                 match request {
-                    RpcRequest::ConfirmedSlot { tx } => {
-                        let result = rpc.get_confirmed_slot().await;
+                    RpcRequest::Slots { tx } => {
+                        let result =
+                            tokio::try_join!(rpc.get_finalized_slot(), rpc.get_confirmed_slot());
                         let _ = tx.send(result);
                     }
                     RpcRequest::Block { slot, tx } => {
@@ -135,30 +186,6 @@ fn handle_rpc(item: Option<RpcRequest>, rpc: &Arc<RpcSource>) -> bool {
         }
         None => {
             error!("RPC requests stream is finished");
-            true
-        }
-    }
-}
-
-async fn handle_stream(
-    item: Option<Result<StreamSourceMessage, RecvError>>,
-    stream_tx: &mpsc::Sender<StreamSourceMessage>,
-) -> bool {
-    match item {
-        Some(Ok(message)) => {
-            if stream_tx.send(message).await.is_err() {
-                error!("failed to send a message to the stream");
-                true
-            } else {
-                false
-            }
-        }
-        Some(Err(error)) => {
-            error!(?error, "gRPC stream error");
-            true
-        }
-        None => {
-            error!("gRPC stream is finished");
             true
         }
     }