+++ CHANGELOG.md
@@ -14,11 +14,16 @@ The minor version will be incremented upon a breaking change and the patch versi
 
 ### Features
 
-- rpc: add smart cache ([#34](https://github.com/lamports-dev/alpamayo/pull/34))
-
 ### Breaking
 
-## [0.4.0] - 2025-04-22
+## [0.5.0] - 2025-04-26
+
+### Features
+
+- rpc: add smart cache ([#34](https://github.com/lamports-dev/alpamayo/pull/34))
+- source: add gRPC reconnect ([#35](https://github.com/lamports-dev/alpamayo/pull/35))
+
+## x[0.4.0] - 2025-04-22
 
 ### Features